refactor message create flow

+++ morny-system-lib/src/main/scala/cc/sukazyo/cono/morny/system/telegram_api/Standardize.scala
@@ -4,7 +4,7 @@ object Standardize {
 	
 	type UserID = Long
 	type ChatID = Long
-	type MessageThreadID = Int
+	type MessageThreadID = Long
 	type MessageID = Int
 	type MessageGroupID = String
